Add the ability use a different identity than the operator SA when deploying to targets.
This improves the security aspects of the code because we can now limit where
the remote secrets can be deployed to on per-namespace basis.

Because of the tests for the new functionality in controllers/client_factory_test.go,
this also updates the controller-runtime to v0.15.0 which caused some other
collateral changes in otherwise unrelated parts of the codebase.